refactor(gh-887): decommission analysis chip rows, replaced by metrics dashboard
- Remove InfoChipRow (and its SpeedChipRow/GeometryChipRow/PolarChipRow/StabilityChipRow)
  from AnalysisViewerPanel; metrics now live in the global MetricsDashboard panel.
- Remove useDesignAssumptions and useRecomputeStatus from AnalysisViewerPanel (no longer
  needed there); update AnalysisWingGate test to drop the now-stale useDesignAssumptions
  mock.
- Delete scratch route app/workbench/metrics-preview/ (click-dummy replaced by live data).
- Extract type definitions (SpeedData, BalanceData, GaugeData, MetricItem, etc.) from
  metricsMock.ts into metricsTypes.ts so app code (MetricsDashboard, primitives,
  metricsAdapters) imports types without pulling in mock data. metricsMock.ts re-exports
  all types for backward-compat test imports.
- Update MetricsDashboard, primitives, and metricsAdapters to import from metricsTypes.
- chip-row component files kept: test suite (InfoChipRow.test.tsx, SpeedChipRow.test.tsx,
  GeometryChipRow.test.tsx, PolarChipRow.test.tsx, StabilityChipRow.test.tsx) still
  references them directly.
- lastRunTime / lastRunDurationMs kept in Props type (callers still pass them) but
  prefixed _unused in destructuring since buildAnalysisRightSlot was removed from the
  component body.

All 1638 frontend unit tests pass (Node 22).
